Seperated UserJobInfo controller into own file
